# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface ErrorResponse {
  error: string;
}

export type SyncStateInputState = { [key: string]: unknown };

export interface SyncStateInput {
  state: SyncStateInputState;
}

export type SyncStateState = { [key: string]: unknown };

export interface SyncState {
  state: SyncStateState;
  updatedAt: string;
}

export interface HealthStatus {
  status: string;
}