# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: sync
    description: Cross-device sync of app state
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /sync/{syncId}:
    parameters:
      - name: syncId
        in: path
        required: true
        description: Random client-generated sync code. Treat it as a secret; anyone holding it can read and overwrite the state.
        schema:
          type: string
          pattern: "^[A-Za-z0-9_-]{16,64}$"
    get:
      operationId: getSyncState
      tags: [sync]
      summary: Get synced app state
      responses:
        "200":
          description: The stored state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncState"
        "404":
          description: Nothing stored for this sync code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: putSyncState
      tags: [sync]
      summary: Replace synced app state (last write wins)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncStateInput"
      responses:
        "200":
          description: The stored state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncState"
        "400":
          description: Invalid sync code or body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SyncStateInput:
      type: object
      properties:
        state:
          type: object
          additionalProperties: true
      required:
        - state
    SyncState:
      type: object
      properties:
        state:
          type: object
          additionalProperties: true
        updatedAt:
          type: string
          format: date-time
      required:
        - state
        - updatedAt
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status