# lib/api-zod/src/generated/types/syncState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SyncStateState } from "./syncStateState";

export interface SyncState {
  state: SyncStateState;
  updatedAt: Date;
}