# lib/api-zod/src/generated/types/syncStateInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SyncStateInputState } from "./syncStateInputState";

export interface SyncStateInput {
  state: SyncStateInputState;
}